# docs/swagger.yaml
basePath: /
definitions:
  github_com_akozadaev_go_es_analytical_system_internal_models.BusinessType:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
    type: object
  github_com_akozadaev_go_es_analytical_system_internal_models.Demographics:
    properties:
      age_group:
        type: string
      average_income:
        type: number
      interests:
        items:
          type: string
        type: array
      population_density:
        type: number
    type: object
  github_com_akozadaev_go_es_analytical_system_internal_models.GeoPoint:
    properties:
      lat:
        description: Широта (latitude)
        type: number
      lon:
        description: Долгота (longitude)
        type: number
    type: object
  github_com_akozadaev_go_es_analytical_system_internal_models.Location:
    properties:
      address:
        type: string
      business_types_suitable:
        items:
          type: string
        type: array
      city:
        type: string
      competition_density:
        type: number
      coordinates:
        $ref: '#/definitions/github_com_akozadaev_go_es_analytical_system_internal_models.GeoPoint'
      created_at:
        type: string
      demographics:
        $ref: '#/definitions/github_com_akozadaev_go_es_analytical_system_internal_models.Demographics'
      description:
        type: string
      embedding:
        items:
          type: number
        type: array
      id:
        type: string
      name:
        type: string
      region:
        type: string
      score:
        description: Для ранжирования
        type: number
      traffic_score:
        type: number
      updated_at:
        type: string
    type: object
  github_com_akozadaev_go_es_analytical_system_internal_models.OllamaAutocompleteRequest:
    properties:
      prefix:
        type: string
    type: object
  github_com_akozadaev_go_es_analytical_system_internal_models.OllamaAutocompleteResponse:
    properties:
      completion:
        type: string
    type: object
  github_com_akozadaev_go_es_analytical_system_internal_models.OllamaChatRequest:
    properties:
      prompt:
        type: string
    type: object
  github_com_akozadaev_go_es_analytical_system_internal_models.OllamaChatResponse:
    properties:
      content:
        type: string
    type: object
  github_com_akozadaev_go_es_analytical_system_internal_models.RecommendRequest:
    properties:
      business_type:
        description: Тип бизнеса (обязательно)
        type: string
      city:
        description: Город для фильтрации (опционально)
        type: string
      limit:
        description: Максимальное количество результатов (по умолчанию 20)
        type: integer
      region:
        description: Регион для поиска (обязательно)
        type: string
    type: object
  github_com_akozadaev_go_es_analytical_system_internal_models.RecommendResponse:
    properties:
      locations:
        items:
          $ref: '#/definitions/github_com_akozadaev_go_es_analytical_system_internal_models.Location'
        type: array
      total:
        type: integer
    type: object
  github_com_akozadaev_go_es_analytical_system_internal_models.Region:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      parent_region_id:
        type: integer
      updated_at:
        type: string
    type: object
  readiness.BuildCheck:
    properties:
      commit:
        type: string
      go:
        type: string
      ok:
        type: boolean
      version:
        type: string
    type: object
  readiness.Checks:
    properties:
      build:
        $ref: '#/definitions/readiness.BuildCheck'
      database:
        $ref: '#/definitions/readiness.DatabaseCheck'
      disk:
        $ref: '#/definitions/readiness.DiskCheck'
      memory:
        $ref: '#/definitions/readiness.MemoryCheck'
    type: object
  readiness.DatabaseCheck:
    properties:
      error:
        type: string
      latency_ms:
        type: number
      ok:
        type: boolean
      timeout_sec:
        type: number
    type: object
  readiness.DiskCheck:
    properties:
      error:
        type: string
      free_bytes:
        type: integer
      min_free_bytes:
        type: integer
      ok:
        type: boolean
      path:
        type: string
      skipped:
        type: string
      total_bytes:
        type: integer
      unsupported_os:
        type: boolean
      used_percent:
        type: number
    type: object
  readiness.MemoryCheck:
    properties:
      error:
        type: string
      ok:
        type: boolean
      process_heap_alloc_bytes:
        type: integer
      process_heap_idle_bytes:
        type: integer
      process_heap_inuse_bytes:
        type: integer
      process_heap_sys_bytes:
        type: integer
      process_num_gc:
        type: integer
      skipped:
        type: string
      system_available_bytes:
        type: integer
      system_total_bytes:
        type: integer
      system_used_bytes:
        type: integer
    type: object
  readiness.Report:
    properties:
      checks:
        $ref: '#/definitions/readiness.Checks'
      status:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: akozadaev@inbox.ru
    name: API Support
    url: https://github.com/akozadaev/go_es_analytical_system
  description: REST API для рекомендательной системы локаций для бизнеса. Система
    предоставляет рекомендации локаций на основе анализа трафика, конкуренции и демографических
    данных.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Location Recommendation System API
  version: "1.0"
paths:
  /business-types:
    get:
      consumes:
      - application/json
      description: Возвращает все доступные типы бизнеса из справочника
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_akozadaev_go_es_analytical_system_internal_models.BusinessType'
            type: array
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить список типов бизнеса
      tags:
      - business-types
  /health:
    get:
      consumes:
      - application/json
      description: Возвращает статус сервиса. Используется для мониторинга и проверки
        доступности.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Проверка работоспособности сервиса
      tags:
      - health
  /locations/{id}:
    get:
      consumes:
      - application/json
      description: Возвращает полную информацию о локации по её идентификатору
      parameters:
      - description: Идентификатор локации
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_akozadaev_go_es_analytical_system_internal_models.Location'
        "404":
          description: Локация не найдена
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить детали локации
      tags:
      - locations
  /locations/recommend:
    post:
      consumes:
      - application/json
      description: Возвращает список рекомендованных локаций для указанного типа бизнеса
        в регионе. Локации ранжируются по релевантности с учетом traffic_score, competition_density
        и демографии.
      parameters:
      - description: Запрос на рекомендации
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_akozadaev_go_es_analytical_system_internal_models.RecommendRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_akozadaev_go_es_analytical_system_internal_models.RecommendResponse'
        "400":
          description: Неверный запрос
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить рекомендации локаций
      tags:
      - locations
  /ollama/autocomplete:
    post:
      consumes:
      - application/json
      description: Запрашивает продолжение по префиксу (ollamaclient.Autocomplete).
      parameters:
      - description: Префикс кода
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_akozadaev_go_es_analytical_system_internal_models.OllamaAutocompleteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_akozadaev_go_es_analytical_system_internal_models.OllamaAutocompleteResponse'
        "400":
          description: Неверный запрос
          schema:
            additionalProperties:
              type: string
            type: object
        "502":
          description: Ollama недоступен или ошибка вызова
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Автодополнение кода через Ollama
      tags:
      - ollama
  /ollama/chat:
    post:
      consumes:
      - application/json
      description: Отправляет промпт в чат-модель (ollamaclient.Chat) и возвращает
        текст ответа.
      parameters:
      - description: Промпт
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_akozadaev_go_es_analytical_system_internal_models.OllamaChatRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_akozadaev_go_es_analytical_system_internal_models.OllamaChatResponse'
        "400":
          description: Неверный запрос
          schema:
            additionalProperties:
              type: string
            type: object
        "502":
          description: Ollama недоступен или ошибка вызова
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Чат через Ollama
      tags:
      - ollama
  /readiness:
    get:
      consumes:
      - application/json
      description: Возвращает статус готовности сервиса, включая проверки БД, диска
        и памяти
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/readiness.Report'
        "503":
          description: Сервис не готов
          schema:
            $ref: '#/definitions/readiness.Report'
      summary: Проверка готовности сервиса
      tags:
      - health
  /regions:
    get:
      consumes:
      - application/json
      description: Возвращает все доступные регионы из справочника с поддержкой иерархии
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_akozadaev_go_es_analytical_system_internal_models.Region'
            type: array
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить список регионов
      tags:
      - regions
schemes:
- http
- https
swagger: "2.0"