# internal/handlers/handlers.go
// Package handlers содержит HTTP обработчики для REST API рекомендательной системы.
package handlers

import (
	"encoding/json"
	"log"
	"net/http"
	"time"

	"github.com/akozadaev/go_es_analytical_system/internal/config"
	"github.com/akozadaev/go_es_analytical_system/internal/models"
	"github.com/akozadaev/go_es_analytical_system/internal/storage"
	readiness "github.com/akozadaev/go_readiness"
	"github.com/gorilla/mux"
	"ollamaclient"
)

// Handlers содержит зависимости для обработки HTTP запросов.
// Использует Elasticsearch для поиска локаций и PostgreSQL для справочников.
type Handlers struct {
	esStorage *storage.ElasticsearchStorage // Хранилище для Elasticsearch/OpenSearch
	pgStorage *storage.PostgresStorage      // Хранилище для PostgreSQL
	cfg       *config.Config                // Конфигурация приложения
	ollama    *ollamaclient.Client          // Клиент Ollama (OpenAI-совместимый API)
}

// NewHandlers создает новый экземпляр Handlers с заданными хранилищами.
func NewHandlers(
	esStorage *storage.ElasticsearchStorage,
	pgStorage *storage.PostgresStorage,
	cfg *config.Config,
	ollama *ollamaclient.Client,
) *Handlers {
	return &Handlers{
		esStorage: esStorage,
		pgStorage: pgStorage,
		cfg:       cfg,
		ollama:    ollama,
	}
}

// RecommendLocations обрабатывает POST запрос на получение рекомендаций локаций.
// Принимает RecommendRequest в теле запроса и возвращает отсортированный список локаций.
// Эндпоинт: POST /locations/recommend
//
// @Summary      Получить рекомендации локаций
// @Description  Возвращает список рекомендованных локаций для указанного типа бизнеса в регионе. Локации ранжируются по релевантности с учетом traffic_score, competition_density и демографии.
// @Tags         locations
// @Accept       json
// @Produce      json
// @Param        request  body      models.RecommendRequest  true  "Запрос на рекомендации"
// @Success      200      {object}  models.RecommendResponse
// @Failure      400      {object}  map[string]string  "Неверный запрос"
// @Failure      500      {object}  map[string]string  "Внутренняя ошибка сервера"
// @Router       /locations/recommend [post]
func (h *Handlers) RecommendLocations(w http.ResponseWriter, r *http.Request) {
	if r.Method != http.MethodPost {
		http.Error(w, "Method not allowed", http.StatusMethodNotAllowed)
		return
	}

	var req models.RecommendRequest
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		http.Error(w, "Invalid request body", http.StatusBadRequest)
		return
	}

	if req.Region == "" || req.BusinessType == "" {
		http.Error(w, "Region and business_type are required", http.StatusBadRequest)
		return
	}

	if req.Limit == 0 {
		req.Limit = 20
	}

	locations, err := h.esStorage.RecommendLocations(r.Context(), &req)
	if err != nil {
		log.Printf("Error recommending locations: %v", err)
		http.Error(w, "Internal server error", http.StatusInternalServerError)
		return
	}

	// Преобразуем указатели в значения для JSON
	locationValues := make([]models.Location, len(locations))
	for i, loc := range locations {
		locationValues[i] = *loc
	}

	response := models.RecommendResponse{
		Locations: locationValues,
		Total:     len(locationValues),
	}

	w.Header().Set("Content-Type", "application/json")
	if err := json.NewEncoder(w).Encode(response); err != nil {
		log.Printf("Error encoding response: %v", err)
		http.Error(w, "Internal server error", http.StatusInternalServerError)
		return
	}
}

// GetLocation обрабатывает GET запрос на получение детальной информации о локации по ID.
// Эндпоинт: GET /locations/{id}
//
// @Summary      Получить детали локации
// @Description  Возвращает полную информацию о локации по её идентификатору
// @Tags         locations
// @Accept       json
// @Produce      json
// @Param        id   path      string  true  "Идентификатор локации"
// @Success      200  {object}  models.Location
// @Failure      404  {object}  map[string]string  "Локация не найдена"
// @Failure      500  {object}  map[string]string  "Внутренняя ошибка сервера"
// @Router       /locations/{id} [get]
func (h *Handlers) GetLocation(w http.ResponseWriter, r *http.Request) {
	if r.Method != http.MethodGet {
		http.Error(w, "Method not allowed", http.StatusMethodNotAllowed)
		return
	}

	vars := mux.Vars(r)
	id := vars["id"]

	if id == "" {
		http.Error(w, "Location ID is required", http.StatusBadRequest)
		return
	}

	location, err := h.esStorage.GetLocation(r.Context(), id)
	if err != nil {
		if err.Error() == "location not found" {
			http.Error(w, "Location not found", http.StatusNotFound)
			return
		}
		log.Printf("Error getting location: %v", err)
		http.Error(w, "Internal server error", http.StatusInternalServerError)
		return
	}

	w.Header().Set("Content-Type", "application/json")
	if err := json.NewEncoder(w).Encode(location); err != nil {
		log.Printf("Error encoding response: %v", err)
		http.Error(w, "Internal server error", http.StatusInternalServerError)
		return
	}
}

// GetBusinessTypes обрабатывает GET запрос на получение списка всех типов бизнеса.
// Возвращает данные из справочника PostgreSQL.
// Эндпоинт: GET /business-types
//
// @Summary      Получить список типов бизнеса
// @Description  Возвращает все доступные типы бизнеса из справочника
// @Tags         business-types
// @Accept       json
// @Produce      json
// @Success      200  {array}   models.BusinessType
// @Failure      500  {object}  map[string]string  "Внутренняя ошибка сервера"
// @Router       /business-types [get]
func (h *Handlers) GetBusinessTypes(w http.ResponseWriter, r *http.Request) {
	if r.Method != http.MethodGet {
		http.Error(w, "Method not allowed", http.StatusMethodNotAllowed)
		return
	}

	businessTypes, err := h.pgStorage.GetBusinessTypes(r.Context())
	if err != nil {
		log.Printf("Error getting business types: %v", err)
		http.Error(w, "Internal server error", http.StatusInternalServerError)
		return
	}

	// Преобразуем указатели в значения для JSON
	btValues := make([]models.BusinessType, len(businessTypes))
	for i, bt := range businessTypes {
		btValues[i] = *bt
	}

	w.Header().Set("Content-Type", "application/json")
	if err := json.NewEncoder(w).Encode(btValues); err != nil {
		log.Printf("Error encoding response: %v", err)
		http.Error(w, "Internal server error", http.StatusInternalServerError)
		return
	}
}

// GetRegions обрабатывает GET запрос на получение списка всех регионов.
// Возвращает данные из справочника PostgreSQL с поддержкой иерархии.
// Эндпоинт: GET /regions
//
// @Summary      Получить список регионов
// @Description  Возвращает все доступные регионы из справочника с поддержкой иерархии
// @Tags         regions
// @Accept       json
// @Produce      json
// @Success      200  {array}   models.Region
// @Failure      500  {object}  map[string]string  "Внутренняя ошибка сервера"
// @Router       /regions [get]
func (h *Handlers) GetRegions(w http.ResponseWriter, r *http.Request) {
	if r.Method != http.MethodGet {
		http.Error(w, "Method not allowed", http.StatusMethodNotAllowed)
		return
	}

	regions, err := h.pgStorage.GetRegions(r.Context())
	if err != nil {
		log.Printf("Error getting regions: %v", err)
		http.Error(w, "Internal server error", http.StatusInternalServerError)
		return
	}

	// Преобразуем указатели в значения для JSON
	regionValues := make([]models.Region, len(regions))
	for i, r := range regions {
		regionValues[i] = *r
	}

	w.Header().Set("Content-Type", "application/json")
	if err := json.NewEncoder(w).Encode(regionValues); err != nil {
		log.Printf("Error encoding response: %v", err)
		http.Error(w, "Internal server error", http.StatusInternalServerError)
		return
	}
}

// OllamaChat обрабатывает POST /ollama/chat — запрос к локальной чат-модели Ollama.
//
// @Summary      Чат через Ollama
// @Description  Отправляет промпт в чат-модель (ollamaclient.Chat) и возвращает текст ответа.
// @Tags         ollama
// @Accept       json
// @Produce      json
// @Param        request  body      models.OllamaChatRequest  true  "Промпт"
// @Success      200      {object}  models.OllamaChatResponse
// @Failure      400      {object}  map[string]string  "Неверный запрос"
// @Failure      502      {object}  map[string]string  "Ollama недоступен или ошибка вызова"
// @Router       /ollama/chat [post]
func (h *Handlers) OllamaChat(w http.ResponseWriter, r *http.Request) {
	if r.Method != http.MethodPost {
		http.Error(w, "Method not allowed", http.StatusMethodNotAllowed)
		return
	}

	var req models.OllamaChatRequest
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		http.Error(w, "Invalid request body", http.StatusBadRequest)
		return
	}
	if req.Prompt == "" {
		http.Error(w, "prompt is required", http.StatusBadRequest)
		return
	}

	content, err := h.ollama.Chat(r.Context(), req.Prompt)
	if err != nil {
		log.Printf("Ollama chat error: %v", err)
		writeOllamaUpstreamError(w, err)
		return
	}

	w.Header().Set("Content-Type", "application/json")
	if err := json.NewEncoder(w).Encode(models.OllamaChatResponse{Content: content}); err != nil {
		log.Printf("Error encoding Ollama chat response: %v", err)
		http.Error(w, "Internal server error", http.StatusInternalServerError)
		return
	}
}

// OllamaAutocomplete обрабатывает POST /ollama/autocomplete — автодополнение кода через Ollama.
//
// @Summary      Автодополнение кода через Ollama
// @Description  Запрашивает продолжение по префиксу (ollamaclient.Autocomplete).
// @Tags         ollama
// @Accept       json
// @Produce      json
// @Param        request  body      models.OllamaAutocompleteRequest  true  "Префикс кода"
// @Success      200      {object}  models.OllamaAutocompleteResponse
// @Failure      400      {object}  map[string]string  "Неверный запрос"
// @Failure      502      {object}  map[string]string  "Ollama недоступен или ошибка вызова"
// @Router       /ollama/autocomplete [post]
func (h *Handlers) OllamaAutocomplete(w http.ResponseWriter, r *http.Request) {
	if r.Method != http.MethodPost {
		http.Error(w, "Method not allowed", http.StatusMethodNotAllowed)
		return
	}

	var req models.OllamaAutocompleteRequest
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		http.Error(w, "Invalid request body", http.StatusBadRequest)
		return
	}
	if req.Prefix == "" {
		http.Error(w, "prefix is required", http.StatusBadRequest)
		return
	}

	completion, err := h.ollama.Autocomplete(r.Context(), req.Prefix)
	if err != nil {
		log.Printf("Ollama autocomplete error: %v", err)
		writeOllamaUpstreamError(w, err)
		return
	}

	w.Header().Set("Content-Type", "application/json")
	if err := json.NewEncoder(w).Encode(models.OllamaAutocompleteResponse{Completion: completion}); err != nil {
		log.Printf("Error encoding Ollama autocomplete response: %v", err)
		http.Error(w, "Internal server error", http.StatusInternalServerError)
		return
	}
}

// GetReadiness обрабатывает GET запрос на проверку готовности сервиса.
// Проверяет подключение к БД, дисковое пространство, память и возвращает статус.
// Эндпоинт: GET /readiness
//
// @Summary      Проверка готовности сервиса
// @Description  Возвращает статус готовности сервиса, включая проверки БД, диска и памяти
// @Tags         health
// @Accept       json
// @Produce      json
// @Success      200  {object}  readiness.Report
// @Failure      503  {object}  readiness.Report  "Сервис не готов"
// @Router       /readiness [get]
func (h *Handlers) GetReadiness(w http.ResponseWriter, r *http.Request) {
	if r.Method != http.MethodGet {
		http.Error(w, "Method not allowed", http.StatusMethodNotAllowed)
		return
	}

	const defaultReadinessDBTimeout = 5 * time.Second

	dbTimeout := time.Duration(h.cfg.ReadinessDBTimeoutSec) * time.Second
	if h.cfg.ReadinessDBTimeoutSec <= 0 {
		dbTimeout = defaultReadinessDBTimeout
	}

	version := h.cfg.BuildVersion
	if version == "" {
		version = "dev"
	}

	mbToBytes := func(mb int) uint64 {
		return uint64(mb) * 1024 * 1024
	}

	opts := readiness.Options{
		DBTimeout:        dbTimeout,
		DiskPath:         h.cfg.ReadinessDiskPath,
		DiskMinFreeBytes: mbToBytes(h.cfg.ReadinessDiskMinFreeMB),
		RAMMinFreeBytes:  mbToBytes(h.cfg.ReadinessRAMMinFreeMB),
		Version:          version,
		Commit:           h.cfg.GitCommit,
	}

	rep, ok := readiness.Run(r.Context(), h.pgStorage.DB(), opts)
	status := http.StatusOK
	if !ok {
		status = http.StatusServiceUnavailable
	}

	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(status)
	if err := json.NewEncoder(w).Encode(rep); err != nil {
		log.Printf("Error encoding readiness response: %v", err)
		http.Error(w, "Internal server error", http.StatusInternalServerError)
		return
	}
}

func writeOllamaUpstreamError(w http.ResponseWriter, err error) {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusBadGateway)
	if encErr := json.NewEncoder(w).Encode(map[string]string{
		"error":   "ollama_request_failed",
		"details": err.Error(),
	}); encErr != nil {
		log.Printf("encode Ollama error response: %v", encErr)
	}
}

// HealthCheck обрабатывает GET запрос на проверку работоспособности сервиса.
// Используется для мониторинга и проверки доступности API.
// Эндпоинт: GET /health
//
// @Summary      Проверка работоспособности сервиса
// @Description  Возвращает статус сервиса. Используется для мониторинга и проверки доступности.
// @Tags         health
// @Accept       json
// @Produce      json
// @Success      200  {object}  map[string]string
// @Router       /health [get]
func (h *Handlers) HealthCheck(w http.ResponseWriter, r *http.Request) {
	w.Header().Set("Content-Type", "application/json")
	json.NewEncoder(w).Encode(map[string]string{
		"status": "ok",
	})
}